requirements.txt
flask
Flask==2.3.2
gunicorn==21.2.0
flask_mysqldb==0.2.0
python-dotenv==1.0.1
pandas
scikit-learn